Ames housing data 2017 to 2022

A dataset containing details on residential sales in Ames since Jan 1 2017. This data is publicly available through the Ames Assessor at https://www.cityofames.org/government/departments-divisions-a-h/city-assessor/sales. The variables are as follows (more detail on the Ames website):

ames

Format

A data frame with 6935 rows and 16 variables:

Parcel ID

character with ID.

Address

property address in Ames, IA.

Style

factor variable detailing the type of housing.

Occupancy

factor variable of type of housing.

Sale Date

date of sale.

Sale Price

sales price (in US dollar).

Multi Sale

logical value: was this sale part of a package?

YearBuilt

integer value: year in which the house was built.

Acres

acres of the lot.

TotalLivingArea (sf)

total living area in square feet.

Bedrooms

number of bedrooms.

FinishedBsmtArea (sf)

total area of the finished basement in square feet.

LotArea(sf)

total lot area in square feet.

AC

logical value: does the property have an AC?

FirePlace

logical value: does the property have an fireplace?

Neighborhood

factor variable - levels indicate neighborhood area in Ames.